+++ sdk/keyvault/azure-keyvault-keys/tests/test_crypto_client_async.py
@@ -22,8 +22,8 @@ import pytest
 
 from _shared.helpers_async  import get_completed_future
 from _shared.json_attribute_matcher import json_attribute_matcher
-from _test_case_async import KeysTestCase
-from _test_case_base import suffixed_test_name
+from _shared.test_case_async import KeyVaultTestCase
+from _test_case import KeysTestCase, suffixed_test_name
 
 KeyVaultPreparer = functools.partial(
     PowerShellPreparer,
@@ -35,7 +35,7 @@ KeyVaultPreparer = functools.partial(
 NO_GET = Permissions(keys=[p.value for p in KeyPermissions if p.value != "get"])
 
 
-class CryptoClientTests(KeysTestCase):
+class CryptoClientTests(KeysTestCase, KeyVaultTestCase):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, match_body=False, custom_request_matchers=[json_attribute_matcher], **kwargs)
